fix test_describe

+++ tests/test_pce_analysis_results.py
@@ -87,7 +87,7 @@ def test_full_results(results):
 
 
 def test_describe(results):
-    assert(results.describe().to_dict() == {
+    assert(results.describe().astype(float).to_dict() == {
         'f': {'10%': 0.009410762945528006,
               '90%': 2.1708835276870935,
               'count': 25.0,
